Streamlined and fixed segmented evaluation
Got rid of unnecessary methods around the metric calculation. Shorter than 30s observations are ignored and an empty dict is returned to avoid errors (which is later filtered out).
Instead of hardcoding eval_trim_beats to be 0 in segmented runs, it should now be manually set when launching compute_paper_metrics.py. compute_item() is unchanged compared to the
original and compute_item_segmented() was changed so that all segmented logic lives only inside it and ThreadPoolExecutor() calls the correct method based on self.hparams.segment_metrics.