Fixed segmented evaluation, separate evals for beats and downbeats are now calculated.
Implemented a filter in compute_predictions() as, in segmented mode, it is possible for the metrics for an observation to be an empty dict. To avoid index mismatches with the other dictionaries
(pred, piece, dataset), these indices are removed from all. We are interested in the metrics, if those can't be calculated, there is no reason to include them. Otherwise, the original formats
are returned for all for items. Simplified main() under if 'len(args.models) == 1:' by essentially returning to the original version of this part of the script. Since now pl_module.py produces
segmented results in the same exact format as in the non-segmented case, there's no reason to make additional changes. This also simplifies the code and logic.