Retrospective wandb update added
Added a CLI argument --update_wandb to compute_paper_metrics.py which expects a wandb run ID to
be updated with newly calculated metrics (currently, segmented F-measure beat and downbeat
predictions are hard coded). Perhaps later a list of metrics could be provided to the script
which are then updated, instead of a fixed set of metrics.